Add files via upload
For this project, you need to take a decimal number length measurement (in feet) that is input into cell B4 of the starter file ("Nearest Eighth of an Inch - STARTER.xlsm" - note that this is a macro-enabled .xlsm file so be sure to enable macros) and convert that measurement to the number of feet, inches, and nearest 1/8th of an inch.  These measurements (feet, inches, nearest 1/8”) will go into cells B7:B9, respectively.

Here’s the thought process for doing this using the measurement 7.4 feet.

See the screencast "Nearest Eighth of an Inch preview" for an example calculation and all requirements for this project.

First, we pluck off the integer part to obtain 7 feet.

Next, we can subtract that integer part from the initial measurement (7.4 – 7) = 0.4 to get the leftover fraction of a foot.

There are 12 inches in a foot, so we can multiply the fraction (0.4) by 12 to get 4.8 inches.  Thus, 7.4 feet is equal to 7 feet plus 4 inches plus some change.

Do a similar thing to step 2 above to obtain the fraction of an inch.  In our example, we would have (4.8 – 4) = 0.8 inches remaining.

Finally, we just need to round this remaining fraction of an inch (0.8) to the nearest 1/8th of an inch.  1/8th of an inch is equal to 0.125 inches.

For international folks who are perhaps unfamiliar with inches and feet as units of measurements, you can learn more 
here
.

The screencasts "Rounding and remainders" and "Examples: Rounding and remainders" in Week 2 of Part 1 of "Everyday Excel" will help with this problem.

When you feel that your solution works well (and your solution updates automatically when the measurement in cell B4 is changed), click on GRADE.  If you are correct, you'll earn the completion code, which you can enter into the "Nearest Eighth of an Inch Submission Quiz" and earn credit for your work.